requirements.txt
python-jenkins>=1.7,<1.8
click>=7.1,<8.0
PyYaml>=5.3.1<6.0
